+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Bug Fixes
 
+* Fixed Handlebars brace escaping in user input, Handlebars no caching option and URL percent encoding #1158.
 * Prevented double request for images during internationalisation #1158.
 * Fixed position of server-generated error messages for simple WInputs #1161.
 * Fixed partial text matches in dropdown typeahead #1164.