+++ CHANGELOG.md
@@ -5,10 +5,15 @@
 ### API Changes
 
 ### Bug Fixes
-* Modified `HtmlSanitizerUtil` and `HtmlToXMLUtil` to handle escaping brackets. `WTextArea` now defaults to
-  `santizeOutput` on #1158.
 
 ### Enhancements
+
+## Release 1.3.11
+
+### Bug Fixes
+
+* Modified `HtmlSanitizerUtil` and `HtmlToXMLUtil` to handle escaping brackets. `WTextArea` now defaults to
+  `santizeOutput` on #1158.
 
 ## Release 1.3.10